wrangle data with Pandas
The aim of this lab is to provide you with a refresher on the Pandas library so that you can pre-process and analyze the datasets before applying data visualization techniques to them. This lab will work as a crash course on pandas. if you are interested in learning more about the panda's library, detailed descriptions and explanation of how to use it and how to clean, munge, and process data stored in a pandas data frame are provided in our course Data Analysis with Python and Python for Applied Data Science